[TAB_NAVIGATION] - Movendo a implementacao da sdk da estimote da activity para o navigationFragment
