--- .idea/libraries/estimote_sdk_.xml
@@ -1,10 +1,0 @@
-<component name="libraryTable">
-  <library name="estimote-sdk-">
-    <CLASSES>
-      <root url="file://$PROJECT_DIR$/app/build/intermediates/exploded-aar/estimote-sdk/res" />
-      <root url="jar://$PROJECT_DIR$/app/build/intermediates/exploded-aar/estimote-sdk/jars/classes.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
