+++ .idea/libraries/sdk_0_8_7.xml
@@ -1,0 +1,10 @@
+<component name="libraryTable">
+  <library name="sdk-0.8.7">
+    <CLASSES>
+      <root url="file://$PROJECT_DIR$/app/build/intermediates/exploded-aar/com.estimote/sdk/0.8.7/res" />
+      <root url="jar://$PROJECT_DIR$/app/build/intermediates/exploded-aar/com.estimote/sdk/0.8.7/jars/classes.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
