+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/NavApp.iml" filepath="$PROJECT_DIR$/NavApp.iml" />
+      <module fileurl="file://$PROJECT_DIR$/NavApp-1.iml" filepath="$PROJECT_DIR$/NavApp-1.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
       <module fileurl="file://$PROJECT_DIR$/volley/volley.iml" filepath="$PROJECT_DIR$/volley/volley.iml" />
     </modules>
