+++ app/src/main/java/com/ciandt/thegarage/navapp/MainActivity.java
@@ -39,30 +39,6 @@ import java.util.concurrent.TimeUnit;
 
 public class MainActivity extends AppCompatActivity {
 
-    private static final String TAG = MainActivity.class.getSimpleName();
-
-    private static final int NOTIFICATION_ID = 123;
-
-    // Y positions are relative to height of bg_distance image.
-    private static final double RELATIVE_START_POS = 320.0 / 1110.0;
-    private static final double RELATIVE_STOP_POS = 885.0 / 1110.0;
-
-    private BeaconManager beaconManager;
-
-    private Beacon beacon;
-    private Region region;
-
-    private int startY = -1;
-    private int segmentLength = -1;
-
-    private int beaconsSize = 0;
-    private ArrayList<Beacon> beacons;
-
-    private static final int REQUEST_ENABLE_BT = 1234;
-    private static final Region ALL_ESTIMOTE_BEACONS_REGION = new Region("rid", null, null, null);
-    private Repository repository;
-    private TimerRun timerRun;
-
     private ViewPager mViewPager;
     private Context mContext;
     private MainPagerAdapter mMainPagerAdapter;
@@ -71,10 +47,6 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-
-        beaconManager = new BeaconManager(this);
-        timerRun = new TimerRun(this);
-        repository = new Repository(this);
 
         mViewPager = (ViewPager) findViewById(R.id.view_pager_main);
         mMainPagerAdapter = new MainPagerAdapter(getSupportFragmentManager(), MainActivity.this);
@@ -92,42 +64,8 @@ public class MainActivity extends AppCompatActivity {
                 return getResources().getColor(R.color.white);
             }
         });
+
         slidingTabLayout.setViewPager(mViewPager);
-
-        beaconManager.setForegroundScanPeriod(TimeUnit.SECONDS.toMillis(Constants.PERIOD_MILLIS_SCAN_RANGING), Constants.WAIT_TIME_MILLIS_SCAN_RANGING);
-
-        // CallBack Scan Beacons
-        beaconManager.setRangingListener(new BeaconManager.RangingListener() {
-            @Override
-            public void onBeaconsDiscovered(Region region, final List<Beacon> beacons) {
-                runOnUiThread(new Runnable() {
-                    @Override
-                    public void run() {
-
-                        Log.i(TAG, "setRangingListener--onBeaconsDiscovered");
-                        JSONObject jsonObj = new JSONObject();
-
-                        try {
-
-                            Log.i(TAG, "setRangingListener--onBeaconsDiscovered::" + beacons.toString());
-
-                            if(beacons != null && beacons.size() > 0){
-                                Toast.makeText(getApplicationContext(), "macAddress:" + beacons.get(0).getMacAddress().toString(), Toast.LENGTH_LONG).show();
-
-                                jsonObj.put("macAddress", beacons.get(0).getMacAddress());
-                                jsonObj.put("rssi", beacons.get(0).getRssi());
-
-                                repository.put(Constants.BEACON_ANTERIOR_KEY, repository.get(Constants.BEACON_ATUAL_KEY));
-                                repository.put(Constants.BEACON_ATUAL_KEY, jsonObj.toString());
-                            }
-
-                        } catch (JSONException e) {
-                            e.printStackTrace();
-                        }
-                    }
-                });
-            }
-        });
     }
 
     @Override
@@ -148,27 +86,11 @@ public class MainActivity extends AppCompatActivity {
     }
 
     @Override protected void onDestroy() {
-        beaconManager.disconnect();
         super.onDestroy();
     }
 
     @Override protected void onStart() {
         super.onStart();
-
-        //Verificando se o Device tem Bluetooth
-        if (!beaconManager.hasBluetooth()) {
-            Toast.makeText(this, "O aplicativo não irá funcionar, pois seu Celular não tem Bluetooth.", Toast.LENGTH_LONG).show();
-            return;
-        }
-
-        //Verificando se o Bluetooth está habilitando
-        if (!beaconManager.isBluetoothEnabled()) {
-            Intent enableBtIntent = new Intent(BluetoothAdapter.ACTION_REQUEST_ENABLE);
-            startActivityForResult(enableBtIntent, REQUEST_ENABLE_BT);
-        } else {
-            connectToServiceScannBeacon();
-            //timerRun.start();
-        }
     }
 
     @Override protected void onResume(){
@@ -177,45 +99,6 @@ public class MainActivity extends AppCompatActivity {
     }
 
     @Override protected void onStop() {
-        /*try {
-            //Stop busca de Beacon
-            beaconManager.stopRanging(ALL_ESTIMOTE_BEACONS_REGION);
-        } catch (RemoteException e) {
-            Log.d(TAG, "Error while stopping ranging", e);
-        }*/
         super.onStop();
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-
-        // Verificando acesso a permissão
-        if (requestCode == REQUEST_ENABLE_BT) {
-            if (resultCode != RESULT_OK) {
-                Toast.makeText(MainActivity.this, "O aplicativo não irá funcionar, pois não foi dada a permissão de habilitar o Bluetooth.", Toast.LENGTH_LONG).show();
-                return;
-            }
-
-            // Aceito
-            if (resultCode == RESULT_OK) {
-                connectToServiceScannBeacon();
-                //timerRun.start();
-            }
-        }
-    }
-
-    private void connectToServiceScannBeacon() {
-        Collections.<Beacon>emptyList();
-        beaconManager.connect(new BeaconManager.ServiceReadyCallback() {
-            @Override
-            public void onServiceReady() {
-                try {
-                    beaconManager.startRanging(ALL_ESTIMOTE_BEACONS_REGION);
-                } catch (RemoteException e) {
-                    Toast.makeText(MainActivity.this, "Cannot start ranging, something terrible happened", Toast.LENGTH_LONG).show();
-                    Log.e(TAG, "Cannot start ranging", e);
-                }
-            }
-        });
     }
 }
