+++ local.properties
@@ -7,5 +7,6 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Tue Aug 25 10:43:33 BRT 2015
-sdk.dir=/Users/alans/Library/Android/sdk
+#Tue Sep 22 19:48:09 BRT 2015
+ndk.dir=/Users/carlos/Library/Android/sdk/ndk-bundle
+sdk.dir=/Users/carlos/Library/Android/sdk
